Move CAN API related functions to canDriver.c / canDriver.h
Specific CAN API functions moved to canDriver files:
* canDriver.c
* canDriver.h

Added description to functions
