+++ NG474-BOOT/Core/Src/main.c
@@ -25,6 +25,7 @@
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
 #include "bootloader.h"
+#include "canDriver.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -34,7 +35,7 @@
 
 /* Private define ------------------------------------------------------------*/
 /* USER CODE BEGIN PD */
-#define FLASH_USER_START_ADDR   ADDR_FLASH_PAGE_128   /* FOR TESTING WE USE BANK2*/
+
 /* USER CODE END PD */
 
 /* Private macro -------------------------------------------------------------*/
@@ -45,40 +46,13 @@
 /* Private variables ---------------------------------------------------------*/
 
 /* USER CODE BEGIN PV */
-extern FDCAN_HandleTypeDef hfdcan1;
 
-/* Global Variable */
-FDCAN_FilterTypeDef sFilterConfig;
-FDCAN_TxHeaderTypeDef TxHeader;
-FDCAN_RxHeaderTypeDef RxHeader;
-
-uint8_t TxData[8] = {0x10, 0x32, 0x54, 0x76, 0x98, 0x00, 0x11, 0x22};
-uint8_t page_complete_ack[1] = {0x00};
-uint8_t write_complete_ack[1] = {0xFF};
-uint8_t RxData[8];
-
-/* Test Data to be programmed */
-  uint64_t Data64_to_write[32] = {
-  0x0000000000000000, 0x1111111111111111, 0x2222222222222222, 0x3333333333333333,
-  0x4444444444444444, 0x5555555555555555, 0x6666666666666666, 0x7777777777777777,
-  0x8888888888888888, 0x9999999999999999, 0xAAAAAAAAAAAAAAAA, 0xBBBBBBBBBBBBBBBB,
-  0xCCCCCCCCCCCCCCCC, 0xDDDDDDDDDDDDDDDD, 0xEEEEEEEEEEEEEEEE, 0xFFFFFFFFFFFFFFFF,
-  0x0011001100110011, 0x2233223322332233, 0x4455445544554455, 0x6677667766776677,
-  0x8899889988998899, 0xAABBAABBAABBAABB, 0xCCDDCCDDCCDDCCDD, 0xEEFFEEFFEEFFEEFF,
-  0x2200220022002200, 0x3311331133113311, 0x6644664466446644, 0x7755775577557755,
-  0xAA88AA88AA88AA88, 0xBB99BB99BB99BB99, 0xEECCEECCEECCEECC, 0xFFDDFFDDFFDDFFDD};
-
-  uint64_t Received_Data64[32];
-  uint32_t received_data_index = 0;
-
-  uint32_t flash_write_address;
 /* USER CODE END PV */
 
 /* Private function prototypes -----------------------------------------------*/
 void SystemClock_Config(void);
 /* USER CODE BEGIN PFP */
-void test_can_init(void);
-uint64_t array_to_uint64(uint8_t *Array);
+
 /* USER CODE END PFP */
 
 /* Private user code ---------------------------------------------------------*/
@@ -117,7 +91,7 @@ int main(void)
   MX_FDCAN1_Init();
   /* USER CODE BEGIN 2 */
   // bootloader_JumpToUserApp();
-  test_can_init();
+  can_init();
   // bootloader_FlashWrite(FLASH_USER_START_ADDR, Data64_to_write);
   /* USER CODE END 2 */
 
@@ -187,112 +161,6 @@ void SystemClock_Config(void)
 
 /* USER CODE BEGIN 4 */
 
-void test_can_init(void)
-{
-	// Configure the FDCAN filters
-	sFilterConfig.IdType = FDCAN_STANDARD_ID;
-	sFilterConfig.FilterIndex = 0;
-	sFilterConfig.FilterType = FDCAN_FILTER_RANGE;
-	sFilterConfig.FilterConfig = FDCAN_FILTER_TO_RXFIFO0;
-	sFilterConfig.FilterID1 = 0x000;
-	sFilterConfig.FilterID2 = 0x7FF;
-	if (HAL_FDCAN_ConfigFilter(&hfdcan1, &sFilterConfig) != HAL_OK)
-	{
-	Error_Handler();
-	}
-
-	// Configure TxHeader
-	TxHeader.Identifier = 0x00;
-	TxHeader.IdType = FDCAN_STANDARD_ID;
-	TxHeader.TxFrameType = FDCAN_DATA_FRAME;
-	TxHeader.DataLength = FDCAN_DLC_BYTES_8;
-	TxHeader.ErrorStateIndicator = FDCAN_ESI_ACTIVE;
-	TxHeader.BitRateSwitch = FDCAN_BRS_OFF;
-	TxHeader.FDFormat = FDCAN_CLASSIC_CAN;
-	TxHeader.TxEventFifoControl = FDCAN_STORE_TX_EVENTS;
-	TxHeader.MessageMarker = 0x00;
-
-	// Activate Interrupt Notifications
-	if (HAL_FDCAN_ActivateNotification(&hfdcan1,
-		  (FDCAN_IT_RX_FIFO0_NEW_MESSAGE), 0) != HAL_OK)
-	{
-		Error_Handler();
-	}
-	// if (HAL_FDCAN_ActivateNotification(&hfdcan1,
-	//		  (FDCAN_IT_TX_FIFO_EMPTY), 0) != HAL_OK)
-	//{
-	//	Error_Handler();
-	//}
-
-	// Start FDCAN controller (continuous listening CAN bus)
-	if (HAL_FDCAN_Start(&hfdcan1) != HAL_OK)
-	{
-		Error_Handler();
-	}
-
-	// Send a Hello message
-	HAL_FDCAN_AddMessageToTxFifoQ(&hfdcan1, &TxHeader, TxData);
-	TxHeader.Identifier = 0x200;
-
-}
-
-void HAL_FDCAN_TxFifoEmptyCallback(FDCAN_HandleTypeDef *hfdcan)
-{
-	// TxHeader.Identifier = 0x200;
-	// HAL_FDCAN_AddMessageToTxFifoQ(&hfdcan1, &TxHeader, TxData);
-}
-
-void HAL_FDCAN_RxFifo0Callback(FDCAN_HandleTypeDef *hfdcan, uint32_t RxFifo0ITs)
-{
-	// Get message
-	HAL_FDCAN_GetRxMessage(&hfdcan1, FDCAN_RX_FIFO0, &RxHeader, RxData);
-
-	// Convert to uint_64
-	Received_Data64[received_data_index] = array_to_uint64(RxData);
-	received_data_index++;
-
-	// Send ACK - resend frames
-	HAL_FDCAN_AddMessageToTxFifoQ(&hfdcan1, &TxHeader, RxData);
-	TxHeader.Identifier += 1u;
-
-	if (received_data_index == 32u){
-		received_data_index = 0;
-		TxHeader.Identifier = 0x300;
-		TxHeader.DataLength = FDCAN_DLC_BYTES_1;
-
-		// Send ACK - Page complete, 32 values received
-		HAL_FDCAN_AddMessageToTxFifoQ(&hfdcan1, &TxHeader, page_complete_ack);
-
-		// Write page
-		bootloader_FlashWrite(FLASH_USER_START_ADDR, Received_Data64);
-
-		// Send ACK - Write page complete
-		TxHeader.Identifier = 0x400;
-		HAL_FDCAN_AddMessageToTxFifoQ(&hfdcan1, &TxHeader, write_complete_ack);
-
-		TxHeader.Identifier = 0x200;
-		TxHeader.DataLength = FDCAN_DLC_BYTES_8;
-	}
-
-}
-
-/**
-  * @brief	Converts an array of 8 uint8_t elements to a uint64_t
-  * @param	Array is an array of 8 uint8_t elements
-  * @retval	uint64_t converted value
-  */
-uint64_t array_to_uint64(uint8_t *Array)
-{
-	  int array_index;
-	  uint64_t Converted_Array = 0;
-	  for (array_index = 0; array_index < 7; array_index++)
-	  {
-		  Converted_Array += Array[array_index];
-		  Converted_Array = (Converted_Array << 8);
-	  }
-	  Converted_Array += Array[7];
-	  return Converted_Array;
-}
 /* USER CODE END 4 */
 
 /**
